More type refactor cleanup
