+++ background/main.ts
@@ -74,10 +74,10 @@ import {
   signedTypedData,
   signedData as signedDataAction,
   signingSliceEmitter,
-  SignDataRequest,
   typedDataRequest,
   signDataRequest,
 } from "./redux-slices/signing"
+import { SignDataRequest } from "./redux-slices/signing/types"
 
 import { SigningMethod, SignTypedDataRequest } from "./services/signing/types"
 import {
