+++ ui/components/SignData/EIP4361Info.tsx
@@ -1,5 +1,5 @@
 import React from "react"
-import { EIP4361Data } from "@tallyho/tally-background/redux-slices/signing"
+import { EIP4361Data } from "@tallyho/tally-background/redux-slices/signing/types"
 import SignDataInfo from "./SignDataInfo"
 
 // can add networks, ideally should come from some sort of network config
